add descriptions to tachiyomi extensions, identifying urls found in sources
This helps with finding out what websites a given multi-source extension
works with.
I'm not sure why this information (or similar) isn't already available
it would be useful to be able to search all extensions for a given website,
or to see which websites an extension like Madara supports.

This change was autogenerated by a script that looks for urls in extensions.
I will probably never update this.