Add tachiyomi-sy metadata and download script
Note that this repo is getting quite large for netlify. I'll be removing the non-extensions from it soon, and adding them to their own repo.

I wonder how hard it would be (or if it would even work) to have github actions build the repo, then delete the files and create redirects to the release download urls.